wblog_version_6

+++ src/main/java/com/ws/web/CommentController.java
@@ -4,6 +4,7 @@ import com.ws.pojo.Comment;
 import com.ws.pojo.User;
 import com.ws.service.BlogService;
 import com.ws.service.CommentService;
+import com.ws.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Controller;
@@ -11,6 +12,7 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.servlet.http.HttpSession;
 
@@ -27,8 +29,13 @@ public class CommentController {
     @Autowired
     private BlogService blogService;
 
+    @Autowired
+    private UserService userService;
+
     @Value("${comment.avatar}")
     private String avatar;
+
+    private static String admin_username = "syen";
 
     @GetMapping("/comments/{blogId}")
     public String comments(@PathVariable Long blogId, Model model){
@@ -60,18 +67,35 @@ public class CommentController {
      */
 
     @PostMapping("/comment_del")
-    public String comment_del(Comment comment, HttpSession session){
+    public String comment_del(Comment comment, HttpSession session, RedirectAttributes attributes){
         Long commentId = comment.getId();
         Long blogId = comment.getBlog().getId();
-       // String commentEmail = comment.getEmail();
-        commentService.setCommentPrant_comment_id(commentId);
-        try{
-            Thread.sleep(1000);
-            commentService.deleteComment(commentId);
-        }catch (Exception e){
-            return "删除失败";
+        String commentEmail = comment.getEmail();
+        String adminEmail = userService.checkUserEmail(admin_username);
+        if(commentEmail.equals(adminEmail)){
+            commentService.setCommentPrant_comment_id(commentId);
+            try{
+                Thread.sleep(1000);
+                commentService.deleteComment(commentId);
+            }catch (Exception e){
+                attributes.addFlashAttribute("delmessage","删除失败");
+                return "redirect:/comments/" + blogId;
+            }
+        }else{
+            if(commentEmail.equals(session.getAttribute("email"))){
+                 commentService.setCommentPrant_comment_id(commentId);
+             try{
+                 Thread.sleep(1000);
+                 commentService.deleteComment(commentId);
+                 }catch (Exception e){
+                 attributes.addFlashAttribute("delmessage","删除失败");
+                       return "redirect:/comments/" + blogId;
+                 }
+            }else {
+                attributes.addFlashAttribute("delmessage","你没有权限");
+            return "redirect:/comments/" + blogId;
+            }
         }
-
         return "redirect:/comments/" + blogId;
     }
